Boeffla Sound: Version 1.1.1 - fix speaker volume control (variable overflow)

+++ sound/soc/codecs/boeffla_sound.c
@@ -1,7 +1,7 @@
 /*
- * Author: andip71, 15.08.2016
+ * Author: andip71
  * 
- * Version 1.0.0
+ * Version 1.1.1
  *
  * This software is licensed under the terms of the GNU General Public
  * License version 2, as published by the Free Software Foundation, and
@@ -17,6 +17,12 @@
 /*
  * Change log:
  * 
+ * 1.1.1 (16.09.2016)
+ *   - Fix speaker control (variable overflow)
+ *
+ * 1.1.0 (23.08.2016)
+ *   - Add speaker control
+ *
  * 1.0.0 (15.08.2016)
  *   - Initial version for OnePlus 3
  * 
@@ -56,7 +62,7 @@ static void reset_boeffla_sound(void)
 
 static void reset_audio_hub(void)
 {
-	u16 tmp;
+	int tmp;
 	
 	// reset all audio hub registers back to defaults
 	set_headphone_gain_l(HEADPHONE_DEFAULT + HEADPHONE_REG_OFFSET);
@@ -185,7 +191,7 @@ static ssize_t headphone_volume_store(struct device *dev, struct device_attribut
 
 static ssize_t speaker_volume_show(struct device *dev, struct device_attribute *attr, char *buf)
 {
-	u16 val;
+	int val;
 
 	val = get_speaker_gain();	// mono speaker, so we alwas treat L and R the same
 	val = (val >> 8) * -1;
@@ -202,7 +208,7 @@ static ssize_t speaker_volume_store(struct device *dev, struct device_attribute 
 	unsigned int ret = -EINVAL;
 	int val;
 	int val_unused;
-	u16 tmp;
+	int tmp;
 
 	// Terminate if boeffla sound is not enabled
 	if (!boeffla_sound)
